// src/models/Category.ts
/**
 * This model represents Category
 */
export type Category = {
    /**
     * Category Id Autogenerated
     */
    id: number;

    /**
     * Category description
     */
    description: string;

    /**
     * Sort order
     */
    sort: number;

    /**
     * Is category active or not
     */
    isActive: boolean;

    /**
     * Id of the user who updated
     */
    updatedBy?: number;

    /**
     * Date and time of last update
     */
    updatedDate?: Date;
}